add option to skip update step in tests

+++ update.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 set -euxo pipefail
+
+# skip git update in testing, so we are testing local changes and not
+# overwriting them with HEAD.
+test -f /srv/kissh/.testing && exit 0
+
 git checkout -f main
 git reset --hard origin/main
 git pull
